# apps/sgd-next-console/src/types/openapi.json
{
  "openapi": "3.1.0",
  "components": {
    "schemas": {
...
      "CategoryInfo1": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "minAge": {
            "type": "integer",
            "format": "int32"
          },
          "maxAge": {
            "type": "integer",
            "format": "int32"
          },
          "ageRange": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "LocalDateTime": {
        "type": "string",
        "format": "date-time",
        "examples": [
          "2022-03-10T12:15:50"
        ]
      },
...
      "SportResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryInfo1"
            }
          },
          "createdAt": {
            "$ref": "#/components/schemas/LocalDateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/LocalDateTime"
          }
        }
      },
      "SportUpdateRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 255,
            "pattern": "\\S"
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT token obtenido a través de Keycloak. El token debe incluir los roles necesarios (ADMIN_GENERAL, ADMIN_CLUB, PROFESOR) y los claims personalizados (club_id, venue_ids, sport_ids).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "SecurityScheme": {
        "type": "openIdConnect",
        "openIdConnectUrl": "http://localhost:8089/realms/SGD/.well-known/openid-configuration",
        "description": "Authentication"
      }
    }
  },
  "info": {
    "contact": {
      "email": "dev@sgd.com",
      "name": "SGD Development Team",
      "url": "https://sgd.com/support"
    },
    "description": "Sistema de Gesti�n Deportiva - API Backend",
    "license": {
      "name": "Proprietary",
      "url": "https://sgd.com/license"
    },
    "title": "SGD Backend API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.sgd.com",
      "description": "Servidor de producción"
    }
  ],
...
  "paths": {
...
    "/api/sports/{id}": {
      "put": {
        "summary": "Actualizar deporte existente",
        "description": "Actualiza los datos de un deporte existente. Solo disponible para ADMIN_GENERAL.",
        "tags": [
          "Sports"
        ],
        "parameters": [
          {
            "description": "ID único del deporte",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SportUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Deporte actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "401": {
            "description": "Token de acceso inválido o expirado"
          },
          "403": {
            "description": "Sin permisos para actualizar deportes"
          },
          "404": {
            "description": "Deporte no encontrado"
          },
          "409": {
            "description": "Ya existe un deporte con ese nombre"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Obtener deporte específico",
        "description": "Obtiene los detalles completos de un deporte incluyendo todas sus categorías activas.",
        "tags": [
          "Sports"
        ],
        "parameters": [
          {
            "description": "ID único del deporte",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deporte encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SportResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos para ver este deporte"
          },
          "404": {
            "description": "Deporte no encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Eliminar deporte",
        "description": "Elimina un deporte del sistema (soft delete). Solo disponible para ADMIN_GENERAL. No se puede eliminar si tiene categorías activas.",
        "tags": [
          "Sports"
        ],
        "parameters": [
          {
            "description": "ID único del deporte",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deporte eliminado exitosamente"
          },
          "400": {
            "description": "No se puede eliminar el deporte (tiene categorías activas)"
          },
          "401": {
            "description": "Token de acceso inválido o expirado"
          },
          "403": {
            "description": "Sin permisos para eliminar deportes"
          },
          "404": {
            "description": "Deporte no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}